Finished Integrating JWT into RestAPI. Still Need Some Testing

+++ api/user_routes.py
@@ -1,5 +1,8 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.responses import JSONResponse
+
+from auth.dependencies import activeUserRequired
+from helpers.helper import isAdmin, isAdminOrOwner
 
 from services.user_services import *
 from services.account_services import deleteAccountsDB
@@ -8,7 +11,13 @@ from models.user import User
 userRouter = APIRouter(tags=["Users"])
 
 @userRouter.get("/")
-def get_all_users():
+def get_all_users(currentUser=Depends(activeUserRequired)):
+
+    if not isAdmin(currentUser):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail={"Not Authorized"}
+        )
 
     users = getUsersDB()
 
@@ -18,13 +27,20 @@ def get_all_users():
     )
 
 @userRouter.get("/{userId}")
-def get_user(userId: str):
+def get_user(userId: str, currentUser=Depends(activeUserRequired)):
+
+    if not isAdminOrOwner(userId, currentUser):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail={"Not Authorized"}
+        )
+
     user = getUserDB(userId)
 
     if not user:
-        return JSONResponse(
-            status_code=404,
-            content={"message": "User not found"}
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail={"User not found"}
         )
 
     return JSONResponse(
@@ -33,13 +49,20 @@ def get_user(userId: str):
     )
 
 @userRouter.get("/{userId}/accounts")
-def get_user_accounts(userId: str):
+def get_user_accounts(userId: str, currentUser=Depends(activeUserRequired)):
+
+    if not isAdminOrOwner(userId, currentUser):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail={"Not Authorized"}
+        )
+
     acc = getUserAccDB(userId)
 
     if not acc:
-        return JSONResponse(
-            status_code=404,
-            content={"message": "User not found"}
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail={"User not found"}
         )
 
     return JSONResponse(
@@ -48,10 +71,23 @@ def get_user_accounts(userId: str):
     )
 
 @userRouter.post("/")
-def create_user(user: User):
+def create_user(user: User, currentUser=Depends(activeUserRequired)):
+
+    if not isAdmin(currentUser):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail={"Not Authorized"}
+        )
+
     newUser = createUserDB(
-        user.name, user.password, user.email, user.status
+        user.name, user.password, user.email, user.status, user.auth
     )
+
+    if newUser == 409:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail={"Duplicate Email"}
+        )
 
     if newUser.acknowledged:
         return JSONResponse(
@@ -62,44 +98,74 @@ def create_user(user: User):
             }
         )
     else:
-        return JSONResponse(
-            status_code=500,
-            content={"message": "Server Error"}
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail={"Database Connection Error"}
         )
 
 @userRouter.put("/{userId}")
-def update_user(userId: str, user: User):
+def update_user(userId: str, user: User, currentUser=Depends(activeUserRequired)):
+
+    if not isAdminOrOwner(userId, currentUser):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail={"Not Authorized"}
+        )
+
     updated = updateUserDB(
-        userId, user.name, user.password, user.email, user.status
+        userId, user.name, user.password, user.email, user.status, user.auth
     )
 
     if not updated or updated.matched_count == 0:
-        return JSONResponse(
-            status_code=404,
-            content={"message": "User not found"}
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail={"User not found"}
         )
-    return JSONResponse(
-        status_code=200,
-        content={
-            "acknowledged": updated.acknowledged,
-            "updatedId": userId,
-        }
-    )
+
+    if updated.acknowledged:
+        return JSONResponse(
+            status_code=200,
+            content={
+                "acknowledged": updated.acknowledged,
+                "updatedId": userId,
+            }
+        )
+    
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail={"Database Connection Error"}
+        )
 
 @userRouter.post("/{userId}/deactivate")
-def deactivate_user(userId: str):
+def deactivate_user(userId: str, currentUser=Depends(activeUserRequired)):
+
+    if not isAdminOrOwner(userId, currentUser):
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail={"Not Authorized"}
+        )
+
     inactive = deleteUserDB(userId)
     closed = deleteAccountsDB(userId)
 
     if not inactive or inactive.matched_count == 0:
-        return JSONResponse(
-            status_code=404,
-            content={"message": "User not found"}
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail={"User not found"}
         )
-    return JSONResponse(
-        status_code=200,
-        content={
-            "acknowledged": "User has been deleted",
-            "accountsClosed": closed.modified_count 
-        }
-    )
+
+    if inactive.acknowledged:
+        return JSONResponse(
+            status_code=200,
+            content={
+                "acknowledged": "User has been deleted",
+                "accountsClosed": closed.modified_count 
+            }
+        )
+
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail={"Database Connection Error"}
+        )
